NVIDIA: SAUCE: iommu/arm-smmu-v3: Add mpam helpers to query and set state
To allow an iommu_group to be moved between resctrl groups as if it
were a CPU thread, the mpam driver needs to be able to set the partid
and pmg for the iommu_group.
Use the properties in the STE, as these only apply to one stream.
The MPAM driver also needs to know the maximum partid and pmg
values that the SMMU can generate. This allows it to determine
the system-wide common supported range of values. Add a helper
to return this id register.
Tested-by: Amit Singh Tomar <amitsinght@marvell.com>

Signed-off-by: James Morse <james.morse@arm.com>
(cherry picked from commit 46a241f45ca9b71abf900f31a0c89fcbf24c44c4 https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git)
Signed-off-by: Fenghua Yu <fenghuay@nvidia.com>