NVIDIA: SAUCE: arm_mpam: resctrl: Add support for 'MB' resource
resctrl supports 'MB', as a percentage throttling of traffic somewhere
after the L3. This is the control that mba_sc uses, so ideally the
class chosen should be as close as possible to the counters used for
mba_local.
MB's percentage control can be backed either with the fixed point
fraction MBW_MAX or the bandwidth portion bitmap. Add a helper to convert
to/from percentages. One problem here is the value written is not the
same as the value read back. This is deliberately made visible to
user-space. Another is the MBW_MAX fixed point fraction can't represent
100%. This is going to confuse user-space, so shift everything up taking
all-ones as 100% and zero as the minimum granularity.
CC: Zeng Heng <zengheng4@huawei.com>
Co-developed-by: Dave Martin <Dave.Martin@arm.com>

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>>
(cherry picked from commit a2e2d2ea3861ba5e63b00775438a08019044da0b https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git)
Signed-off-by: Fenghua Yu <fenghuay@nvidia.com>